Mouse Pause: pause on more actions, make pause length variable
This change makes the mouse pause event fire when a mouseup event and a scroll
event occur, in addition to the previous mousemove. This should make the
behavior much more consistent.

In addition, the timeout value has been turned into a variable so that those
interested in having a longer delay can configure the value through the local
variable store.