Delay questing instead of disabling it
This change implements timers for the Main and Side quests. Instead of
disabling those functions when there are no quests to perform, a timer
is set for 1 week, after which the script will be able to check again
to see if there are new quests.

The logic around how questing is approached feels a little convoluted,
so this will likely require some testing to be sure there are no
negative impacts.